# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cadência API
  version: 0.34.0
  description: API do MVP de planejamento adaptativo de ciclismo.
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://cadencia-api.devsaulo.com.br
    description: Produção na VPS Oracle via Cloudflare Tunnel
paths:
...
  /v1/auth/register:
    post:
      summary: Cria uma conta e inicia uma sessão
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { description: Conta criada, com o aceite dos termos registrado }
        '400': { description: Dados inválidos ou aceite dos termos ausente (terms_required) }
        '409': { description: Versão dos termos desatualizada (terms_outdated) }
  /v1/auth/login:
    post:
      summary: Autentica por e-mail e senha
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: Sessão iniciada }
        '401': { description: Credenciais inválidas }
        '429': { description: Muitas tentativas por IP ou para a mesma conta; consulte Retry-After }
...
  /v1/auth/account:
    delete:
      summary: Apaga definitivamente a conta e todos os dados pessoais
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, confirmation]
              properties:
                password:
                  type: string
                  format: password
                  description: Senha atual da conta
                confirmation:
                  type: string
                  enum: [ENCERRAR CONTA]
                  description: Confirmação explícita exigida pela operação
      responses:
        '204': { description: Conta e dados pessoais apagados }
        '400': { description: Confirmação ou dados inválidos }
        '401': { description: Sessão ausente ou senha incorreta }
        '500': { description: Falha sem alteração dos dados }
  /v1/auth/account/export:
    get:
      summary: Baixa uma planilha com os dados essenciais do usuário autenticado (LGPD, art. 18)
      description: >
        Arquivo .xlsx com três abas: Conta, Treinos realizados e Atividades importadas.
        Não inclui dados de saúde (limitações, dor, check-ins, medidas corporais) nem
        senha e códigos de sessão; a cópia completa é entregue por solicitação ao contato
        da Política de Privacidade.
      responses:
        '200':
          description: Planilha para download (Content-Disposition attachment)
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        '401': { description: Sessão ausente ou expirada }
        '429': { description: Muitas exportações em pouco tempo }
        '500': { description: Falha ao montar o arquivo }
  /v1/auth/accept-terms:
    post:
      summary: Registra que o usuário aceitou a versão vigente dos Termos de Uso e da Política de Privacidade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [terms_version]
              properties:
                terms_version: { type: string, example: '2026-10-01' }
      responses:
        '204': { description: Aceite registrado }
        '401': { description: Sessão ausente ou expirada }
        '409': { description: Versão dos termos desatualizada (terms_outdated) }
  /v1/auth/change-password:
    post:
      summary: Altera a senha do usuário autenticado e desconecta as demais sessões
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password:
                  type: string
                  format: password
                  minLength: 10
                  maxLength: 72
                  description: Deve ser diferente da senha atual
      responses:
        '200': { description: Senha alterada; a sessão atual é mantida e as demais são revogadas }
        '400': { description: Nova senha inválida ou igual à atual }
        '401': { description: Sessão ausente ou senha atual incorreta }
        '429': { description: Muitas tentativas }
  /v1/auth/logout-others:
    post:
      summary: Encerra todas as sessões do usuário, exceto a atual
      responses:
        '200':
          description: Sessões revogadas
          content:
            application/json:
              schema:
                type: object
                required: [revoked_sessions]
                properties:
                  revoked_sessions: { type: integer, minimum: 0 }
        '401': { description: Sessão ausente ou expirada }
  /v1/auth/resend-verification:
    post:
      summary: Reenvia o link de confirmação para o usuário autenticado
      responses:
        '200': { description: Link enviado ou conta já confirmada }
        '401': { description: Sessão ausente ou expirada }
        '503': { description: Serviço de e-mail indisponível }
  /v1/auth/verify-email:
    post:
      summary: Confirma o endereço de e-mail com um token de uso único
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenRequest' }
      responses:
        '200': { description: E-mail confirmado }
        '400': { description: Token inválido, expirado ou já usado }
  /v1/auth/forgot-password:
    post:
      summary: Solicita um link de redefinição sem revelar se o e-mail possui conta
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailRequest' }
      responses:
        '200': { description: Resposta genérica enviada }
        '503': { description: Serviço de e-mail indisponível }
  /v1/auth/reset-password:
    post:
      summary: Redefine a senha com um token de uso único e revoga as sessões existentes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200': { description: Senha redefinida }
        '400': { description: Senha ou token inválido }
  /v1/me:
    get:
      summary: Retorna o usuário da sessão e se ele aceitou a versão vigente dos termos
      responses:
        '200':
          description: Usuário autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { type: object }
                  legal:
                    type: object
                    properties:
                      terms_version: { type: string, example: '2026-10-01' }
                      accepted: { type: boolean }
        '401': { description: Sessão ausente ou expirada }
  /v1/profile:
    get:
      summary: Retorna o perfil do atleta autenticado
      responses:
        '200': { description: Perfil encontrado ou nulo quando ainda não preenchido }
        '401': { description: Sessão ausente ou expirada }
    put:
      summary: Cria ou atualiza o perfil básico de ciclismo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AthleteProfile' }
      responses:
        '200': { description: Perfil salvo }
        '400': { description: Dados inválidos }
        '401': { description: Sessão ausente ou expirada }
  /v1/onboarding:
    get:
      summary: Retorna limitações, objetivos, disponibilidade e contexto de ciclismo do atleta
      responses:
        '200': { description: Dados do onboarding }
        '401': { description: Sessão ausente ou expirada }
        '409': { description: Perfil básico ainda não preenchido }
  /v1/onboarding/questionnaire:
    get:
      summary: Retorna o contrato versionado de etapas e perguntas condicionais do onboarding de ciclismo
      responses:
        '200':
          description: Questionário autenticado disponível
          content:
            application/json:
              schema:
                type: object
                required: [questionnaire]
                properties:
                  questionnaire: { $ref: '#/components/schemas/Questionnaire' }
        '401': { description: Sessão ausente ou expirada }
  /v1/onboarding/limitations:
    put:
      summary: Substitui as limitações ativas informadas pelo atleta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [limitations]
              properties:
                limitations:
                  type: array
                  maxItems: 10
                  items: { $ref: '#/components/schemas/Limitation' }
      responses:
        '200': { description: Limitações salvas }
        '400': { description: Dados inválidos }
        '401': { description: Sessão ausente ou expirada }
  /v1/onboarding/goals:
    put:
      summary: Substitui os objetivos principal e secundário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [goals]
              properties:
                goals:
                  type: array
                  minItems: 1
                  maxItems: 2
                  items: { $ref: '#/components/schemas/Goal' }
      responses:
        '200': { description: Objetivos salvos }
        '400': { description: Dados inválidos }
        '401': { description: Sessão ausente ou expirada }
  /v1/onboarding/availability:
    put:
      summary: Substitui a disponibilidade dos sete dias da semana
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [availability]
              properties:
                availability:
                  type: array
                  minItems: 7
                  maxItems: 7
                  items: { $ref: '#/components/schemas/Availability' }
      responses:
        '200': { description: Disponibilidade salva }
        '400': { description: Dados inválidos }
        '401': { description: Sessão ausente ou expirada }
  /v1/onboarding/cycling-context:
    put:
      summary: Salva dados opcionais do contexto de ciclismo do atleta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cycling_context]
              properties:
                cycling_context: { $ref: '#/components/schemas/CyclingContext' }
      responses:
        '200': { description: Contexto de ciclismo salvo }
        '400': { description: Dados inválidos }
        '401': { description: Sessão ausente ou expirada }
  /v1/assessments/current:
    get:
      summary: Retorna a avaliação submáxima mais recente do atleta
      responses:
        '200': { description: Avaliação encontrada ou nula }
        '401': { description: Sessão ausente ou expirada }
  /v1/assessments/submaximal:
    post:
      summary: Registra uma avaliação submáxima de ciclismo orientada por RPE
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmaxAssessmentRequest' }
      responses:
        '201': { description: Avaliação registrada }
        '400': { description: Dados inválidos }
        '401': { description: Sessão ausente ou expirada }
  /v1/recovery/today:
    get:
      summary: Retorna o check-in do atleta na data local informada
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: Check-in encontrado ou nulo }
        '400': { description: Data inválida }
        '401': { description: Sessão ausente ou expirada }
    put:
      summary: Salva o check-in diário e pode reduzir a próxima sessão ainda não iniciada
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecoveryRequest' }
      responses:
        '200': { description: Check-in salvo e eventual adaptação informada }
        '400': { description: Dados inválidos }
        '401': { description: Sessão ausente ou expirada }
        '409': { description: Perfil básico ainda não preenchido }
  /v1/evolution/summary:
    get:
      summary: Retorna indicadores históricos observados, incluindo métricas agregadas de pedal quando registradas
      responses:
        '200':
          description: Totais, semanas, comparação entre sessões planejadas e realizadas, métricas agregadas e check-ins de recuperação
          content:
            application/json:
              schema:
                type: object
                required: [summary]
                properties:
                  summary: { $ref: '#/components/schemas/EvolutionSummary' }
        '401': { description: Sessão ausente ou expirada }
  /v1/feedback:
    post:
      summary: Registra o relato de experiência do atleta autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserFeedbackRequest' }
      responses:
        '201': { description: Feedback registrado }
        '400': { description: Categoria, nota ou mensagem inválida }
        '401': { description: Sessão ausente ou expirada }
  /v1/plans/generate:
    post:
      summary: Gera e salva um rascunho de quatro semanas usando regras explicáveis
      responses:
        '201': { description: Plano gerado }
        '401': { description: Sessão ausente ou expirada }
        '409': { description: Onboarding incompleto }
        '403': { description: E-mail ainda não confirmado }
  /v1/plans/current:
    get:
      summary: Retorna o plano ativo ou rascunho mais recente
      responses:
        '200': { description: Plano encontrado ou nulo }
        '401': { description: Sessão ausente ou expirada }
  /v1/activities:
    get:
      summary: Retorna atividades concluídas e canceladas do atleta autenticado, da mais recente para a mais antiga
      responses:
        '200': { description: Histórico de atividades }
        '401': { description: Sessão ausente ou expirada }
  /v1/activities/import:
    post:
      summary: Importa uma atividade a partir de um arquivo .fit ou .gpx (upload multipart, campo "file", limite 20 MB)
      description: Extrai duração, distância, elevação, FC, potência e cadência do arquivo e sugere treinos planejados no mesmo dia. Nunca altera um treino nem o plano; a confirmação é feita pelos endpoints de conclusão/correção do treino.
      responses:
        '201': { description: Atividade importada, com os treinos do mesmo dia sugeridos como vínculo }
        '400': { description: Corpo multipart inválido ou sem o campo "file" }
        '401': { description: Sessão ausente ou expirada }
        '409': { description: Este arquivo já foi importado por este atleta }
        '413': { description: Arquivo maior que o limite de 20 MB }
        '415': { description: Extensão não reconhecida (envie .fit ou .gpx) }
        '422': { description: Arquivo corrompido, sem dados de atividade ou de um esporte diferente de ciclismo }
  /v1/activities/imported:
    get:
      summary: Lista as atividades importadas do atleta autenticado, da mais recente para a mais antiga
      responses:
        '200': { description: Atividades importadas }
        '401': { description: Sessão ausente ou expirada }
  /v1/activities/imported/{activityID}:
    delete:
      summary: Remove uma atividade importada (não afeta nenhum treino ou plano)
      parameters:
        - in: path
          name: activityID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Atividade removida }
        '401': { description: Sessão ausente ou expirada }
        '404': { description: Atividade não encontrada ou pertencente a outro atleta }
  /v1/activities/imported/{activityID}/workout:
    put:
      summary: Vincula uma atividade importada a um treino do atleta, ou remove o vínculo
      description: Grava só a associação entre a atividade e o treino, para consulta. Não copia dados para o treino, não altera o plano e não passa pelo motor; a execução registrada continua vindo dos endpoints de conclusão/correção. Envie workout_id nulo para desvincular.
      parameters:
        - in: path
          name: activityID
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [workout_id]
              properties:
                workout_id: { type: string, format: uuid, nullable: true }
      responses:
        '200': { description: "Atividade com o vínculo atualizado: workout_id, workout_name e workout_scheduled_on quando vinculada" }
        '400': { description: Corpo inválido ou identificador de treino malformado }
        '401': { description: Sessão ausente ou expirada }
        '404': { description: Atividade ou treino não encontrados para este atleta }
  /v1/activities/imported/{activityID}/candidates:
    get:
      summary: Sugere treinos planejados para uma atividade já importada
      description: Considera o dia da atividade e os dias vizinhos, porque a data local só era conhecida no upload.
      parameters:
        - in: path
          name: activityID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: "Lista candidate_workouts, com id, scheduled_on, name e status de cada treino" }
        '401': { description: Sessão ausente ou expirada }
        '404': { description: Atividade não encontrada ou pertencente a outro atleta }
  /v1/plans/{planID}/activate:
    post:
      summary: Ativa um rascunho e cancela o plano ativo anterior do atleta
      parameters:
        - in: path
          name: planID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Plano ativado }
        '400': { description: Identificador inválido }
        '401': { description: Sessão ausente ou expirada }
        '404': { description: Rascunho não encontrado ou pertencente a outro atleta }
        '403': { description: E-mail ainda não confirmado }
  /v1/protection/recovered:
    post:
      summary: "Registra a declaração 'Estou recuperado' e reavalia os treinos futuros planejados"
      description: "Só está disponível com a proteção graduada ligada (PROTECTION_LEVELS_ENABLED). A declaração vale como um check-in bom: rebaixa uma proteção que já teve tempo de diminuir, mas não uma proteção forte por dor relatada nos últimos dias. Repetir no mesmo dia não duplica o registro."
      responses:
        '200': { description: Plano atual após a reavaliação }
        '401': { description: Sessão ausente ou expirada }
        '404': { description: Recurso indisponível ou plano inexistente }
        '409': { description: Perfil incompleto }
  /v1/workouts/{workoutID}/start:
    post:
      summary: Inicia um treino planejado pertencente ao plano ativo
      parameters:
        - $ref: '#/components/parameters/WorkoutID'
      responses:
        '200': { description: Sessão iniciada e plano atualizado }
        '400': { description: Identificador inválido }
        '401': { description: Sessão de usuário ausente ou expirada }
        '404': { description: Treino não encontrado no plano ativo }
        '409': { description: Transição de estado inválida }
  /v1/workouts/{workoutID}/explanation:
    post:
      summary: Gera uma explicação em linguagem simples sem alterar a prescrição do treino
      parameters:
        - $ref: '#/components/parameters/WorkoutID'
      responses:
        '200': { description: Explicação gerada ou fallback validado pelas regras quando a IA estiver indisponível }
        '400': { description: Identificador inválido }
        '401': { description: Sessão de usuário ausente ou expirada }
        '404': { description: Treino não encontrado no plano atual }
  /v1/workouts/{workoutID}/complete:
    post:
      summary: Conclui uma sessão em andamento e salva o feedback percebido
      parameters:
        - $ref: '#/components/parameters/WorkoutID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteWorkoutRequest' }
      responses:
        '200': { description: Sessão concluída, feedback salvo e plano atualizado }
        '400': { description: Identificador ou feedback inválido }
        '401': { description: Sessão de usuário ausente ou expirada }
        '404': { description: Treino não encontrado no plano ativo }
        '409': { description: Treino não está em andamento }
  /v1/workouts/{workoutID}/correct:
    post:
      summary: Corrige métricas opcionais de uma sessão concluída inelegível
      description: Substitui distância, elevação, frequência cardíaca e potência opcionais. Null remove a métrica; duração, RPE, feedback e prescrição não são alterados. O valor anterior fica preservado para auditoria.
      parameters:
        - $ref: '#/components/parameters/WorkoutID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkoutCorrectionRequest' }
      responses:
        '200': { description: Métricas corrigidas, integridade reavaliada e plano preservado }
        '400': { description: Identificador ou métrica inválida }
        '401': { description: Sessão de usuário ausente ou expirada }
        '404': { description: Treino não encontrado no plano do atleta }
        '409': { description: Sessão não concluída ou sem inconsistência corrigível }
  /v1/workouts/{workoutID}/cancel:
    post:
      summary: Cancela uma sessão em andamento e marca o treino como não realizado
      parameters:
        - $ref: '#/components/parameters/WorkoutID'
      responses:
        '200': { description: Sessão cancelada e plano atualizado }
        '400': { description: Identificador inválido }
        '401': { description: Sessão de usuário ausente ou expirada }
        '404': { description: Treino não encontrado no plano ativo }
        '409': { description: Treino não está em andamento }
  /v1/workouts/{workoutID}/missed:
    post:
      summary: Marca um treino passado como não realizado sem criar uma sessão
      parameters:
        - $ref: '#/components/parameters/WorkoutID'
      responses:
        '200': { description: Treino marcado como não realizado e plano atualizado }
        '400': { description: Identificador inválido }
        '401': { description: Sessão de usuário ausente ou expirada }
        '404': { description: Treino não encontrado no plano ativo }
        '409': { description: Treino ainda não venceu ou já foi iniciado, concluído ou encerrado }
components:
  parameters:
    WorkoutID:
      in: path
      name: workoutID
      required: true
      schema: { type: string, format: uuid }
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, display_name, accept_terms, terms_version]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 10, maxLength: 72 }
        display_name: { type: string, minLength: 2, maxLength: 100 }
        accept_terms:
          type: boolean
          description: Deve ser true; o cadastro é recusado sem o aceite dos Termos de Uso e da Política de Privacidade
        terms_version:
          type: string
          description: Versão dos textos que o usuário viu (AAAA-MM-DD); precisa ser a versão vigente
          example: '2026-10-01'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    EmailRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    TokenRequest:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 1 }
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token: { type: string, minLength: 1 }
        password: { type: string, minLength: 10, maxLength: 72 }
    AthleteProfile:
      type: object
      required: [experience_level]
      properties:
        birth_date: { type: [string, 'null'], format: date }
        sex:
          type: [string, 'null']
          enum: [female, male, other, prefer_not_to_say, null]
        height_cm: { type: [number, 'null'], minimum: 100, maximum: 250 }
        weight_kg: { type: [number, 'null'], minimum: 30, maximum: 350 }
        waist_cm: { type: [number, 'null'], minimum: 30, maximum: 250 }
        body_fat_percent: { type: [number, 'null'], minimum: 2, maximum: 70 }
        weight_trend: { type: string, enum: [not_informed, stable, increasing, decreasing] }
        experience_level: { type: string, enum: [beginner, intermediate, advanced] }
        activity_level: { type: [string, 'null'] }
    Limitation:
      type: object
      required: [kind, description, is_active, professional_clearance_recommended]
      properties:
        kind: { type: string, enum: [pain, injury, medical_condition, mobility, other] }
        description: { type: string, minLength: 3, maxLength: 500 }
        location: { type: string, maxLength: 120 }
        intensity: { type: [integer, 'null'], minimum: 1, maximum: 10, description: Escala percebida opcional; não é diagnóstico. }
        aggravating_movement: { type: string, maxLength: 200 }
        started_on: { type: [string, 'null'], format: date }
        symptoms_during_after:
          type: array
          maxItems: 5
          uniqueItems: true
          items: { type: string, enum: [dizziness, unusual_shortness_of_breath, malaise, extreme_fatigue, other] }
        medical_restriction: { type: boolean }
        recent_surgery: { type: boolean }
        exercise_prohibited: { type: boolean }
        condition_affecting_exercise: { type: boolean }
        is_active: { type: boolean }
        professional_clearance_recommended: { type: boolean }
    Goal:
      type: object
      required: [goal_type, priority, details]
      properties:
        goal_type: { type: string, enum: [health, fitness, endurance, performance, event, weight_management] }
        priority: { type: integer, minimum: 1, maximum: 2 }
        target_date: { type: [string, 'null'], format: date }
        details: { type: string, maxLength: 500 }
    Availability:
      type: object
      required: [weekday, available_minutes]
      properties:
        weekday: { type: integer, minimum: 0, maximum: 6 }
        available_minutes: { type: integer, minimum: 0, maximum: 480 }
        preferred_time: { type: [string, 'null'], pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        location: { type: [string, 'null'], maxLength: 80 }
    Questionnaire:
      type: object
      required: [version, scope, steps, questions]
      properties:
        version: { type: string, example: cycling-onboarding-v2 }
        scope: { type: string, example: cycling_profile_and_plan_generation }
        steps:
          type: array
          minItems: 4
          items: { $ref: '#/components/schemas/QuestionnaireStep' }
        questions:
          type: array
          items: { $ref: '#/components/schemas/QuestionDefinition' }
...
    SubmaxAssessmentRequest:
      type: object
      required: [duration_minutes, actual_rpe, pain_reported]
      properties:
        duration_minutes: { type: integer, minimum: 15, maximum: 30 }
        actual_rpe: { type: number, minimum: 1, maximum: 10 }
        pain_reported: { type: boolean }
        notes: { type: string, maxLength: 1000 }
...